[🔧수정] 서당 페이지 axiosInstance 적용
