+++ src/axios/school/korea/ku_1202.ts
@@ -1,6 +1,6 @@
 /* 고려대학교 - 사용자 맞춤 알약 검색 시스템 1202 */
-import axios from 'axios'
 import base64DataToFile from '../../base64DataToFile'
+import axiosInstance from '@/services/axiosInstance'
 
 const ku1202 = async (
   value: any, // 사용자가 입력한 값 (string or base64)
@@ -8,7 +8,7 @@ const ku1202 = async (
   setLoading: any, // 로딩
   // setResult: any,    // 결과 컴포넌트
 ) => {
-  const axiosUrl = 'http://aihunmin-edu.t3q.ai:8181/api/inference/files_req_ajx' // 고정값
+  const axiosUrl = '/api/inference/files_req_ajx' // 고정값
 
   /* image가 2장이 들어옴 */
   const imageA = value[0] // 배열에 담긴 첫번째 이미지
@@ -28,7 +28,7 @@ const ku1202 = async (
 
   /* axios 비동기 통신 함수 */
   try {
-    const res = await axios.post(axiosUrl, formData, {
+    const res = await axiosInstance.post(axiosUrl, formData, {
       headers: {
         'Content-Type': 'multipart/form-data',
       },
