+++ src/axios/school/korea/ku_1203.ts
@@ -1,6 +1,6 @@
 /* 고려대학교 - 가구 상세 분류 1203 */
-import axios from 'axios'
 import base64DataToFile from '../../base64DataToFile'
+import axiosInstance from '@/services/axiosInstance'
 
 const ku1203 = async (
   value: any, // 사용자가 입력한 값 (string or base64)
@@ -16,7 +16,7 @@ const ku1203 = async (
     '다리 없음',
     '바형 다리',
   ]
-  const axiosUrl = 'http://aihunmin-edu.t3q.ai:8181/api/inference/file_req_ajx' // 고정값
+  const axiosUrl = '/api/inference/file_req_ajx' // 고정값
   const convertData = await base64DataToFile(value, 'image', 'image/png')
   /* FormData (apiUrl, data) 형태로 전송 */
   const formData = new FormData()
@@ -28,7 +28,7 @@ const ku1203 = async (
 
   /* axios 비동기 통신 함수 */
   try {
-    const res = await axios.post(axiosUrl, formData, {
+    const res = await axiosInstance.post(axiosUrl, formData, {
       headers: {
         'Content-Type': 'multipart/form-data',
       },
