+++ src/axios/school/kyungpook/knu_1100.ts
@@ -1,14 +1,14 @@
-/* 고려대학교 - 보행약자 보조 AI 1205 */
-import axios from 'axios'
+/* 경북대학교 - 이륜차 위험요소 탐지 1100 */
 import base64DataToFile from '../../base64DataToFile'
+import axiosInstance from '@/services/axiosInstance'
 
-const ku1205 = async (
+const knu1100 = async (
   value: any, // 사용자가 입력한 값 (string or base64)
   formUrl: any, // 사용자가 입력한 api Url
   setLoading: any, // 로딩
   // setResult: any,    // 결과 컴포넌트
 ) => {
-  const axiosUrl = 'http://aihunmin-edu.t3q.ai:8181/api/inference/file_req_ajx' // 고정값
+  const axiosUrl = '/api/inference/file_req_ajx' // 고정값
   const convertData = await base64DataToFile(value, 'image', 'image/jpeg')
   /* FormData (apiUrl, data) 형태로 전송 */
   const formData = new FormData()
@@ -20,7 +20,7 @@ const ku1205 = async (
 
   /* axios 비동기 통신 함수 */
   try {
-    const res = await axios.post(axiosUrl, formData, {
+    const res = await axiosInstance.post(axiosUrl, formData, {
       headers: {
         'Content-Type': 'multipart/form-data',
       },
@@ -44,4 +44,4 @@ const ku1205 = async (
   return resultData
 }
 
-export default ku1205
+export default knu1100
