+++ src/axios/school/kyungpook/knu_1102.ts
@@ -1,6 +1,6 @@
 /* 경북대학교 - 버스 승객 이상행동 감지 1102 */
-import axios from 'axios'
 import base64DataToFile from '../../base64DataToFile'
+import axiosInstance from '@/services/axiosInstance'
 
 const knu1102 = async (
   value: any, // 사용자가 입력한 값 (string or base64)
@@ -8,7 +8,7 @@ const knu1102 = async (
   setLoading: any, // 로딩
   // setResult: any,    // 결과 컴포넌트
 ) => {
-  const axiosUrl = 'http://aihunmin-edu.t3q.ai:8181/api/inference/file_req_ajx' // 고정값
+  const axiosUrl = '/api/inference/file_req_ajx' // 고정값
   const convertData = await base64DataToFile(value, 'video', 'video/mp4')
   /* FormData (apiUrl, data) 형태로 전송 */
   const formData = new FormData()
@@ -20,7 +20,7 @@ const knu1102 = async (
 
   /* axios 비동기 통신 함수 */
   try {
-    const res = await axios.post(axiosUrl, formData, {
+    const res = await axiosInstance.post(axiosUrl, formData, {
       headers: {
         'Content-Type': 'multipart/form-data',
       },
